[refactor] Move action and API specific logic into those modules
