+++ Cargo.toml
@@ -14,14 +14,12 @@ resolver = "2"
 [workspace.dependencies]
 anyhow = "1.0.81"
 axum = "0.7.5"
-chrono = "0.4.38"
-dotenv = "0.15.0"
+base64 = "0.22.1"
 futures = "0.3.30"
 hyper = "1.3.1"
-regex-lite = "0.1.6"
+reqwest = {version = "0.12.4", features = ["json"]}
 serde = "1.0.202"
 serde_json = "1.0.117"
-serde_path_to_error = "0.1.16"
 shared = {path = "shared"}
 time = {version = "0.3.36", features = ["local-offset"]}
 tokio = {version = "1.38.0", features = ["rt-multi-thread", "macros"]}
