+++ action/Cargo.toml
@@ -5,9 +5,15 @@ version = "0.1.0"
 
 [dependencies]
 anyhow.workspace = true
-dotenv.workspace = true
+base64.workspace = true
 futures.workspace = true
-regex-lite.workspace = true
+git2 = "0.18.3"
+glob = "0.3.1"
+regex-lite = "0.1.6"
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+serde_yaml = "0.9.34"
 shared.workspace = true
 tokio.workspace = true
 tracing-subscriber.workspace = true
