+++ shared/Cargo.toml
@@ -6,22 +6,16 @@ version = "0.1.0"
 [dependencies]
 anyhow.workspace = true
 axum.workspace = true
-base64 = "0.22.1"
-chrono.workspace = true
-dotenv.workspace = true
+base64.workspace = true
+chrono = "0.4.38"
+dotenv = "0.15.0"
 futures.workspace = true
-git2 = "0.18.3"
-glob = "0.3.1"
 hex = "0.4.3"
 hyper.workspace = true
 jsonwebtoken = "9.3.0"
 openssl = {version = "0.10.64", features = ["vendored"]}
-regex-lite.workspace = true
-reqwest = {version = "0.12.4", features = ["json"]}
+reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-serde_path_to_error.workspace = true
-serde_yaml = "0.9.34"
 tokio.workspace = true
 tracing.workspace = true
-uuid = {version = "1.11.0", features = ["v4"]}
