--- shared/src/ai/mod.rs
@@ -1,7 +1,0 @@
-mod chat_gpt;
-mod claude;
-mod prompts;
-mod schemas;
-
-pub use chat_gpt::{ChatGpt, ReleaseNotes, TestCases};
-pub use claude::Claude;
