+++ shared/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod ai;
 pub mod services;
 pub mod utils;
