+++ shared/src/services/mod.rs
@@ -1,6 +1,4 @@
-pub mod git;
+pub mod chat_gpt;
+pub mod claude;
 pub mod github;
 pub mod jira;
-pub mod slack;
-
-pub use git::Git;
