+++ shared/src/utils/mod.rs
@@ -1,3 +1,2 @@
 pub mod config;
 pub mod error;
-pub mod jwt;
